refactor(chrome): consolidate title-bar menu into sidebar footer
- Strip the SearchTrigger (middle of the title bar) and the AppMenu
  (top-right cog) from the title bar. Command palette is reachable
  via keyboard shortcut + the new sidebar menu; the title bar is
  now just sidebar toggle + resource monitor + IDE launcher +
  window controls.
- Repurpose the sidebar Settings cog into the full app menu (acts
  as the dropdown trigger). Dropdown items: Settings, Command
  palette, Keyboard shortcuts, Documentation, Report issue,
  version info, Sign out — same surface area as the old top-right
  cog, plus shortcut hints on Settings and Command palette.
- Align the tab bar's bottom border with the sidebar separator.
  h-[52px] -> h-[45px] (was 1px below the sidebar's 44px content
  + 1px separator stack) so the horizontal rule reads as a single
  continuous stroke across the window.
